pyproject.toml
[project]
name = "codeyak"
version = "0.0.1"
description = "A code review agent optimized for following strict guidelines"
readme = "README.md"
authors = [
    { name = "rai", email = "raiselmartinez@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.0",
    "gitpython>=3.1.0",
    "instructor>=1.14.1",
    "langfuse>=3.11.2",
    "openai>=2.15.0",
    "pydantic-settings>=2.12.0",
    "python-gitlab>=7.1.0",
    "pyyaml>=6.0.1",
    "tomli-w>=1.0.0",
]

[project.scripts]
yak = "codeyak.apps.cli:main"

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/codeyak"]

# Ensure prebuilt guidelines are included in the package
[tool.hatch.build.targets.wheel.force-include]
"src/codeyak/prebuilt" = "codeyak/prebuilt"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
